Fixes #18
On startup, existing sub-dimensions of personal dimensions were being
given the wrong worldprovider by the launch code.  The reason for this is
that subdimensions are not added to the player -> dimension mapping for
personal dimensions, because there is only one canonical "personal"
dimension.  The launch code was using presence or absence in the mapping
to determine whether a dimension is personal, rather than the dimension
type.  This is now fixed.