Fixes #35
If there is a problem reading the blacklist file from disk, previously,
null would be returned from the DDSaveHandler.readBlacklist method.  This
would result in a crash later on down the line when we tried to iterate
the blacklist.  Now we return an empty array.  I'd like to be able to
diagnose the issues causing the blacklist file to be corrupted, but
unfortunately when this issue crops up, people only post the crash :(